pyproject.toml
[tool.poetry]
name = "bookmarker"
version = "0.1.0"
description = ""
authors = ["Alexander Herrmann <106409+dArignac@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
pyjwt = "^2.10.0"
daphne = "^4.1.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
django-environ = "^0.11.2"
django-oauth-toolkit = "^3.0.1"
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
